Parameterized
Went and parameterized all the functions. Added a few lines at the end
so that I can run this one script on a schedule, and execute it for
myself and my girlfriend.

For anyone who wants to execute this script for themselves, they'll need
to generate a config file and edit line 133.